// src/theme/components/modules/HeroBanner/fields.type.ts
// THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
// Removing the above comment will disable type generation for this module
// This file was created by @hubspot/cms-dev-server, for more information see https://github.hubspot.com/cms-react/reference/js-modules.html#module-fields
import { type TextFieldType, type RequiredField, type FieldGroupWithChildren, type RichTextFieldType, type LinkFieldType, type DefaultValues } from "@hubspot/cms-components/fields";
type HeroBannerFieldsType = DefaultValues<{
    heading: FieldGroupWithChildren<{
        text: RequiredField<TextFieldType>;
        size: "sm" | "md" | "lg" | "xl" | "2xl";
        variant: "regular" | "medium" | "semibold" | "bold";
    }>;
    bannerContent: FieldGroupWithChildren<{
        richTextContentHTML?: RichTextFieldType;
    }>;
    actions: FieldGroupWithChildren<{
        primary: FieldGroupWithChildren<{
            text: RequiredField<TextFieldType>;
            href: RequiredField<LinkFieldType>;
            variant: "primary" | "secondary" | "link";
            size: "sm" | "md" | "lg" | "xl" | "2xl";
        }>;
        secondary: FieldGroupWithChildren<{
            text: RequiredField<TextFieldType>;
            href: RequiredField<LinkFieldType>;
            variant: "primary" | "secondary" | "link";
            size: "sm" | "md" | "lg" | "xl" | "2xl";
        }>;
    }>;
}>;
export default HeroBannerFieldsType;